Reorganize momentary and latched checkboxes
Separated the interference checkboxes into two groups:
momentary - will immediately be re-set after the event is captured to
database.  In other words, these are states that only last a moment and
are self-correcting.
latched - stay set until the state changes
Added an Other interference in the latched category, for things I
haven't identified.